+++ apps/governance/src/components/governance/common/TitleContainer.tsx
@@ -2,7 +2,11 @@
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/apps/blob/main/LICENSE)
 
 const TitleContainer = ({ title }: { title: string }) => {
-  return <div className="text-lg font-bold text-pearl">{title}</div>;
+  return (
+    <div className="text-lg font-bold text-pearl" data-testid="proposal-title">
+      {title}
+    </div>
+  );
 };
 
 export default TitleContainer;