+++ packages/eslint-config-custom/shared.js
@@ -12,16 +12,27 @@ module.exports = {
     "plugin:@typescript-eslint/recommended-requiring-type-checking",
     "plugin:security/recommended",
     "plugin:sonarjs/recommended",
-    "plugin:vitest/recommended"
+    "plugin:vitest/recommended",
   ],
-  plugins: ["@typescript-eslint", "prettier", "sonarjs", "no-secrets", "vitest"],
+  plugins: [
+    "@typescript-eslint",
+    "prettier",
+    "sonarjs",
+    "no-secrets",
+    "vitest",
+  ],
   parser: "@typescript-eslint/parser",
   parserOptions: {
     tsconfigRootDir: ".",
     project: ["./tsconfig.json"],
   },
   rules: {
-    "prettier/prettier": "error",
+    "prettier/prettier": [
+      "error",
+      {
+        endOfLine: "auto",
+      },
+    ],
     "no-unused-vars": "off",
     "@typescript-eslint/restrict-template-expressions": "error",
     "@typescript-eslint/no-unused-vars": ["error"],
@@ -31,7 +42,7 @@ module.exports = {
         checksVoidReturn: false,
       },
     ],
-    "no-secrets/no-secrets":["error",{"tolerance": 4.1}],
+    "no-secrets/no-secrets": ["error", { tolerance: 4.1 }],
     "sonarjs/prefer-single-boolean-return": "off",
     "sonarjs/prefer-immediate-return": "off",
   },