+++ packages/playwright-config-custom/index.ts
@@ -1,0 +1,6 @@
+// @ts-ignore
+const sharedConfig = require("./playwright.shared.config.ts");
+
+module.exports = {
+  sharedConfig,
+};